fixed a large number of bugs affecting the planned trajectory goal

+++ modular_drl_env/util/pybullet_util.py
@@ -339,7 +339,7 @@ class pybullet_util:
             maxNumIterations=max_iterations,
             residualThreshold=threshold
         )
-        return np.float32(joints)
+        return np.float64(joints)
     
     ########
     # misc #
